Modifications to avoid the following compiler warnings:
zoltan_low.c(150): warning #186: pointless comparison of unsigned integer with zero
      assert(nbor_global_id[ii] >= 0);
      ^
zoltan_low.c(215): warning #167: argument of type "char (*)[12]" is incompatible with parameter of type "char **"
    Zoltan_Initialize(1, &argv, &version);
                         ^
cubic_mesh_low.c(137): warning #266: function "printf" declared implicitly
    printf("aa %d %d %d\n", aa[0]->nx, aa[0]->ny, aa[0]->nz);
    ^
cubic_mesh_fft.c(35): warning #266: function "printf" declared implicitly
    printf("Init fourier %d %d %d\n", this[0]->nx, this[0]->ny, this[0]->nz);
    ^
fortcom: Warning: mpi.F90, line 102: A dummy argument with an explicit INTENT(OUT) declaration is not given an explicit value.   [MPI_GRP_OUT]
  subroutine mpi_grp_copy_equal(mpi_grp_out, mpi_grp_in)
write_iter_low.c(150): warning #181: argument is incompatible with corresponding format string conversion
  	sprintf(w->buf+w->pos, "%8u%20.12e", w->iter, w->iter*w->dt); w->pos+=8+20;
 	                                     ^

+++ src/basic/write_iter_low.c
@@ -147,7 +147,7 @@ void FC_FUNC_(write_iter_start, WRITE_ITER_START)
 {
 	write_iter *w=(write_iter *)*v;
 	write_iter_realloc(w, 8+20);
-	sprintf(w->buf+w->pos, "%8u%20.12e", w->iter, w->iter*w->dt); w->pos+=8+20;
+	sprintf(w->buf+w->pos, "%8u%20.12e", (unsigned)(w->iter), w->iter*w->dt); w->pos+=8+20;
 	w->iter++;
 }
